feat: Update the user import script to fail fast
Update the user import script to do more checks on the data file, and fail as soon as possible in order to allow admins to update it and retry.
The script now checks if an email was duplicated, if a user doesn't have a team, and if a team name was written in multiple variations.

Related to #5375